requirements.txt
langchain_community
python-dotenv
streamlit
langchain_core